+++ src/algo.c
@@ -90,7 +90,7 @@ int swap_dyn(dyndata_t* a, dyndata_t* b)
 /*  O : NULL if error                                       */
 /*      Address of the element otherwise                    */
 /************************************************************/
-void* get_arrayelem(meta_t* meta, uint64_t i)
+void* get_arrayelem(meta_t* meta, uint32_t i)
 {
     if(i >= meta->nbelements)
     {
@@ -112,7 +112,7 @@ void* get_arrayelem(meta_t* meta, uint64_t i)
 /*  O : -1 if error                                         */
 /*      0 otherwise                                         */
 /************************************************************/
-int set_arrayelem(meta_t* meta, uint64_t i, void* elem)
+int set_arrayelem(meta_t* meta, uint32_t i, void* elem)
 {
     if(i >= meta->nbelements)
     {
@@ -162,7 +162,7 @@ int listToArray(meta_t* dList, meta_t* dArray, e_listtoarray action){
 
     //copy elements one by one in the array
     tmp_list = dList->structure;
-    for(uint64_t i=0 ; i<dArray->nbelements ; i++){
+    for(uint32_t i=0 ; i<dArray->nbelements ; i++){
         //position the pointer properly
         tmp_array = get_arrayelem(dArray, i);
         memcpy(tmp_array, tmp_list->data, dList->elementsize);
@@ -198,7 +198,7 @@ int listToArray(meta_t* dList, meta_t* dArray, e_listtoarray action){
 /************************************************************/
 int arrayToList(meta_t* dArray, meta_t* dList, e_listtoarray action){
     //copy elements one by one in the list
-    for(uint64_t i=0 ; i<dArray->nbelements ; i++)
+    for(uint32_t i=0 ; i<dArray->nbelements ; i++)
     {
         //insert in the list
         if(insertListSorted(dList,  get_arrayelem(dArray, i)) < 0)
@@ -234,7 +234,7 @@ int arrayToAVL(meta_t* dArray, meta_t* dAVL, e_listtoarray action){
     dyndata_t* tmp_array = dArray->structure;
 
     //copy elements one by one in the list
-    for(uint64_t i=0 ; i<dArray->nbelements ; i++){
+    for(uint32_t i=0 ; i<dArray->nbelements ; i++){
         tmp_array = get_arrayelem(dArray, i);
         //insert in the AVL
         dAVL->structure = insertAVL(dAVL, dAVL->structure, tmp_array);
@@ -258,7 +258,7 @@ int arrayToAVL(meta_t* dArray, meta_t* dAVL, e_listtoarray action){
 /*  O :  0 -> Sorted                                        */
 /*      -1 -> Error                                         */
 /************************************************************/
-int bubbleSortArray(meta_t *meta, uint64_t nb){
+int bubbleSortArray(meta_t *meta, uint32_t nb){
     void *current=NULL, *next=NULL;
     void* tmp = NULL;
 
@@ -285,8 +285,8 @@ int bubbleSortArray(meta_t *meta, uint64_t nb){
         return -1;
     }
 
-    for(uint64_t i=0 ; i<nb ; i++){
-        for(uint64_t j=0 ; j<meta->nbelements-i-1 ; j++){
+    for(uint32_t i=0 ; i<nb ; i++){
+        for(uint32_t j=0 ; j<meta->nbelements-i-1 ; j++){
             //properly place the cursors
             current = get_arrayelem(meta, j);
             next = get_arrayelem(meta, j+1);
@@ -313,10 +313,10 @@ int bubbleSortArray(meta_t *meta, uint64_t nb){
 /*  O :  0 -> Sorted                                        */
 /*      -1 -> Error                                         */
 /************************************************************/
-int bubbleSortList(meta_t* meta, uint64_t nb){
+int bubbleSortList(meta_t* meta, uint32_t nb){
     dyndata_t *current=NULL, *next=NULL, *right_ptr=NULL;
     int swapped = 0;
-    uint64_t count = 0;
+    uint32_t count = 0;
 
     //no meta data available
     if(!meta || !meta->doCompare)
@@ -375,9 +375,9 @@ int bubbleSortList(meta_t* meta, uint64_t nb){
 /************************************************************/
 /*  WARNING : is solely to be used by the quick sort func.! */
 /************************************************************/
-uint64_t quickSortPartitioning(meta_t* meta, uint64_t low, uint64_t high){
+uint32_t quickSortPartitioning(meta_t* meta, uint32_t low, uint32_t high){
     void* pivotElem = NULL, *elem_i=NULL, *elem_j=NULL, *tmp = NULL;
-    uint64_t i = 0;
+    uint32_t i = 0;
 
     //get the element at the highest index of the partition for the pivot
     //      and place i below the lowest
@@ -396,7 +396,7 @@ uint64_t quickSortPartitioning(meta_t* meta, uint64_t low, uint64_t high){
 
     //swap the elements until the pivot is at the right place
     //      with lower elements before, and higher ones after
-    for(uint64_t j=low ; j<=high-1 ; j++){
+    for(uint32_t j=low ; j<=high-1 ; j++){
         elem_j = get_arrayelem(meta, j);
         if((*meta->doCompare)(elem_j, pivotElem) < 0){
             i++;
@@ -427,8 +427,8 @@ uint64_t quickSortPartitioning(meta_t* meta, uint64_t low, uint64_t high){
 /*  O :  0 -> Sorted                                        */
 /*      -1 -> Error                                         */
 /************************************************************/
-int quickSortArray(meta_t* meta, uint64_t low, uint64_t high){
-    uint64_t pivot=0;
+int quickSortArray(meta_t* meta, uint32_t low, uint32_t high){
+    uint32_t pivot=0;
 
     //no meta data available
     if(!meta || !meta->doCompare)
@@ -514,10 +514,10 @@ int binarySearchArray(meta_t *meta, void* toSearch, e_search scope){
 /*  O : Element if found                                    */
 /*      NULL otherwise                                      */
 /************************************************************/
-void* get_listelem(meta_t* meta, uint64_t i)
+void* get_listelem(meta_t* meta, uint32_t i)
 {
     dyndata_t *tmp = meta->structure, *next = NULL;
-    uint64_t index = 0;
+    uint32_t index = 0;
 
     if(i>=meta->nbelements)
     {
@@ -826,7 +826,7 @@ int foreachList(meta_t* meta, void* parameter, int (*doAction)(void*, void*)){
 int foreachArray(meta_t* meta, void* parameter, int (*doAction)(void*, void*)){
     void* tmp = NULL;
 
-    for(uint64_t i=0 ; i<meta->nbelements ; i++){
+    for(uint32_t i=0 ; i<meta->nbelements ; i++){
         //position the pointer properly
         tmp = get_arrayelem(meta, i);
         //execute action
